test(bulk-model-sync-gradle): refactor tests
Extract side effect from test case into ChangeApplier and avoid companion objects in the tests.
